requirements.txt
coverage
dj-database-url
dj-static
Django
django-coverage
django-debug-toolbar
django-dotenv
django-extensions
django-filebrowser-no-grappelli
django-js-asset
django-localflavor
django-mptt
django-nose
django-sitetree
easy-thumbnails
factory-boy
Faker
flake8
mccabe
nose
pep8
Pillow
psycopg2-binary
pycodestyle
pyflakes
python-dateutil
pytz
six
sqlparse
static3
text-unidecode
typing
